requirements.txt
requests
sqlite-vss
python-docx
PyMuPDF
tkinter
numpy 